sha2: New module containing all SHA2 hashes
