+++ src/hazardous/kdf/pbkdf2.rs
@@ -68,7 +68,7 @@
 use crate::{
     errors::UnknownCryptoError,
     hazardous::{
-        hash::sha512::{SHA512_BLOCKSIZE, SHA512_OUTSIZE},
+        hash::sha2::sha512::{SHA512_BLOCKSIZE, SHA512_OUTSIZE},
         mac::hmac,
     },
     util,
