+++ src/typedefs.rs
@@ -657,7 +657,7 @@ macro_rules! construct_hmac_key {
             #[must_use = "SECURITY WARNING: Ignoring a Result can have real security implications."]
             /// Construct from a given byte slice.
             pub fn from_slice(slice: &[u8]) -> Result<$name, UnknownCryptoError> {
-                use crate::hazardous::hash::sha512::{Sha512, SHA512_OUTSIZE};
+                use crate::hazardous::hash::sha2::sha512::{Sha512, SHA512_OUTSIZE};
 
                 let mut secret_key = [0u8; $size];
