+++ tests/hash/mod.rs
@@ -1,11 +1,11 @@
 pub mod blake2b_kat;
 pub mod other_blake2b;
-pub mod sha512_nist_cavp;
 pub mod sha256_nist_cavp;
 pub mod sha384_nist_cavp;
+pub mod sha512_nist_cavp;
 
-use orion::hazardous::hash::{blake2b, sha512, sha256, sha384};
 use crate::TestCaseReader;
+use orion::hazardous::hash::{blake2b, sha2::sha256, sha2::sha384, sha2::sha512};
 
 fn blake2b_test_runner(input: &[u8], key: &[u8], output: &[u8]) {
     // Only make SecretKey if test case key value is not empty.
